AttrMinMax changed from int to float

+++ src/Main.java
@@ -18,7 +18,7 @@ public class Main {
     Hashtable<String,Flight> flights=TapasDataReader.Readers.readFlightDelaysFromDecisions(fnDecisions,steps);
     System.out.println(flights.get("EDDK-LEPA-EWG598-20190801083100").delays[2]);
     Hashtable<String,Vector<Record>> records=TapasDataReader.Readers.readFlightPlans(fnFlightPlans,flights);
-    Hashtable<String,int[]> attrs=new Hashtable<String, int[]>();
+    Hashtable<String,float[]> attrs=new Hashtable<String, float[]>();
     TapasDataReader.Readers.readExplanations(path,steps,flights,attrs);
     for (String s:attrs.keySet())
       System.out.println(s+": ["+attrs.get(s)[0]+".."+attrs.get(s)[1]+"]");
